requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google-auth==2.10.0
google-auth-oauthlib==0.5.2
gspread==5.4.0
idna==3.3
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.11
